Modified code to extend Adafruit_AS7431 class
