--- ESP32/include/AS7341.h
@@ -1,9 +1,0 @@
-#include <Arduino.h>
-#include <Wire.h>
-#include <SPI.h>
-#include <FS.h>
-#include <Adafruit_AS7341.h>
-
-bool initializeSensor(Adafruit_AS7341 &as7341);
-bool setGain(Adafruit_AS7341 &as7341, float gain);
-int getGain(Adafruit_AS7341 &as7341);
