+++ ESP32/include/Serial_Processing.h
@@ -1,4 +1,3 @@
-#include <Adafruit_GFX.h>
 #include <Adafruit_AS7341.h>
 
 unsigned int getSerialIntArgument();
@@ -10,3 +9,5 @@ float getSerialFloatArgument();
 void processSerialBuffer(Adafruit_AS7341 &as7341);
 
 void read_SERIAL(Adafruit_AS7341 &as7341);
+
+bool serialPrintBasicCounts(Stream &serialport, float_t counts[12]);
