+++ ESP32/src/Serial_Processing.cpp
@@ -4,8 +4,7 @@
 
 #include <Led_Control.h>
 #include <Serial_Processing.h>
-#include <AS7341.h>
-#include <Serial_AS7341.h>
+#include <Smartclamp_AS7341.h>
 
 const uint8_t SERIAL_BUFFER_LEN = 128;
 char serialBuffer[SERIAL_BUFFER_LEN];
@@ -25,7 +24,7 @@ float getSerialFloatArgument(){
   return atof(serialBuffer+(bufferPos+1) );
 }
 
-void processSerialBuffer(Adafruit_AS7341 &as7341){
+void processSerialBuffer(Smartclamp_AS7341 &as7341){
   if( toupper(serialBuffer[bufferPos]) == 'L'){
     if( toupper(serialBuffer[bufferPos+1]) == 'O'){
       if( toupper(serialBuffer[bufferPos+2]) == 'N'){
@@ -69,15 +68,7 @@ void processSerialBuffer(Adafruit_AS7341 &as7341){
     if( toupper(serialBuffer[bufferPos+1]) == 'S'){
       if( toupper(serialBuffer[bufferPos+2]) == 'P'){
         // GSP - Get Sensor Parameters
-        Serial.print("Sensor Gain: ");
-        Serial.println(getGain(as7341));
-        Serial.print("Sensor ATime: ");
-        Serial.println(as7341.getATIME());
-        Serial.print("Sensor AStep: ");
-        Serial.println(as7341.getASTEP());
-        Serial.print("Sensor Integration Time (ms): ");
-        Serial.println( (as7341.getATIME()+1) * (as7341.getASTEP()+1) * 2.78 * 0.001);
-        Serial.println();
+        as7341.printParameters(Serial);
       }
     }
   }
@@ -102,4 +93,32 @@ void read_SERIAL(Adafruit_AS7341 &as7341){
     else
       bufferEnd = 0;
   }
+}
+
+bool serialPrintBasicCounts(Stream &serialport, float_t counts[12]){
+    
+    serialport.print("F1 415nm : ");
+    serialport.println(counts[0]);
+    serialport.print("F2 445nm : ");
+    serialport.println(counts[1]);
+    serialport.print("F3 480nm : ");
+    serialport.println(counts[2]);
+    serialport.print("F4 515nm : ");
+    serialport.println(counts[3]);
+    serialport.print("F5 555nm : ");
+    // We skip the duplicates  
+    serialport.println(counts[6]);
+    serialport.print("F6 590nm : ");
+    serialport.println(counts[7]);
+    serialport.print("F7 630nm : ");
+    serialport.println(counts[8]);
+    serialport.print("F8 680nm : ");
+    serialport.println(counts[9]);
+    serialport.print("Clear    : ");
+    serialport.println(counts[10]);
+    serialport.print("NIR      : ");
+    serialport.println(counts[11]);
+
+    serialport.println();
+    return true;
 }
